+++ src/augmenting/AugmentCategoryDisplay.tsx
@@ -1,40 +1,16 @@
 import { groupBy } from "lodash"
 import { useCallback } from "react"
 import { SingleAugmentDisplay } from "./SingleAugmentDisplay"
-import {
-  allAugmentCategories,
-  Augment,
-  augmentByCategory,
-  AugmentCategory,
-  augmentFufillsRequirement,
-  AugmentStat,
-} from "./data/augment"
+import { allAugmentCategories, AugmentCategory } from "./data/augment"
 import { MultiAugmentDisplay } from "./MultiAugmentDisplay"
 import { Box, Paper, Stack, Tab, Tabs } from "@mui/material"
 import { SearchInput } from "components/SearchInput"
-import { atom, useAtom } from "jotai"
-import { atomWithReset } from "jotai/utils"
-
-const augmentCategoryStateAtom = atom<AugmentCategory>(allAugmentCategories[0])
-const searchStatAtom = atomWithReset<AugmentStat | null>(null)
-const searchNameAtom = atomWithReset("")
-
-const availableAugments = atom<Augment[]>((get) => {
-  const category = get(augmentCategoryStateAtom)
-  const searchStat = get(searchStatAtom)
-  const searchName = get(searchNameAtom).toLocaleLowerCase()
-  let filteredAugments = [...augmentByCategory[category]]
-  if (searchName !== "") {
-    filteredAugments = filteredAugments.filter((a) =>
-      a.name.toLocaleLowerCase().includes(searchName),
-    )
-  }
-  if (searchStat) {
-    filteredAugments.filter((a) => augmentFufillsRequirement(a, searchStat))
-  }
-
-  return filteredAugments
-})
+import { useAtom } from "jotai"
+import {
+  availableAugments,
+  augmentCategoryStateAtom,
+  searchNameAtom,
+} from "./augmentSearchState"
 
 const CategoryPane = () => {
   const [augments] = useAtom(availableAugments)
